refactor(tests): move mocks to jest-global-mocks file
